requirements.txt
agentops==0.1.6
crewai==0.28.8
langchain==0.1.16
langchain_community==0.0.34
langchain_groq==0.1.3
langchain_openai==0.1.3
python-dotenv==1.0.1
Requests==2.31.0
streamlit==1.33.0
